+++ src/index.jsx
@@ -4,11 +4,16 @@ import React from 'react';
 import { render } from 'react-dom';
 import { Provider } from 'react-redux';
 
-import App from './components/App.jsx';
-import store from './store/index';
+import App from './App.jsx';
+import store from './app/store';
 
 const root = document.createElement('div');
 root.setAttribute('id', 'root');
 document.body.appendChild(root);
 
-render(<Provider store={store}><App /></Provider>, root);
+render(
+  <Provider store={store}>
+    <App />
+  </Provider>,
+  root,
+);
