+++ webpack.config.js
@@ -4,7 +4,7 @@ const path = require('path');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 
 module.exports = {
-  entry: './client/index.jsx',
+  entry: './src/index.jsx',
   module: {
     rules: [
       {
@@ -33,7 +33,7 @@ module.exports = {
     ],
   },
   output: {
-    path: path.resolve(__dirname, 'client/dist'),
+    path: path.resolve(__dirname, './dist'),
     filename: 'index_bundle.js',
   },
   plugins: [
